refactor(darwin-vz): simplify vmnet helper flows
